Fix cli parsing in packaged app
Yargs automatically strips away process.argv[1] when running in
Electron, because when starting the Electron app via npm/yarn,
process.argv[1] is not the script name, but rather just '.'.
However, when the app is packaged, the '.' is missing and
process.argv[1] will contain the first cli option to parse.

As a workaround, the app is now preprocessing process.argv itself
instead of letting yargs do it in an inconsistent way.