Group animations steps
Instead of having a long list of small animation steps, we switch to a
list of big animation steps that are lists of small animation steps
themselves. Animation steps that belong to the same transformation
action are animated as a block, which makes a lot more sense for
navigating through the transformation actions.