Fix broken full-screen mode on macOS when -f cli option is not present
On macOS, setting the browser windows 'fullscreen' option to 'false'
would not just start the app in windows mode but also disables the
full-screen mode, such that the window can be maximized but not switched
to full-screen.

This fix sets the browser windows 'fullscreen' option to 'true' when
'-f' is provided on the command line but does not set it at all if '-f'
is not supplied.