Add pre-recording facilities to audio recorder
If is often desirable to keep track of a small amount of audio data
before the recording is actually started. This can be useful in
situations where the recording is triggered based on the current volume
level. E.g. if a person starts to speak, the target volume would only be
reached after some (short) period of time and the increase of volume
would be lost leading to an abrupt start of the recording.

During the pre-recording, new audio data will not be appended as usual
but previous data is shifted (partially out of the buffer) to make room
for the new data.