Drop implementation of global timestep cursor
The initial idea was to have a cursor the user can slide over all
elements of the visualization. However, it became apparent that a single
timestep in the neural network output does not directly correspond to
the same timestep in the input. Results for a timestep are based on
several timesteps in the spectrogram. Having such a cursor would
therefore give a false impression of the involved internal mechanisms.