Implement network output shortening and reordering
The output of the neural network may contain tokens that never occured
in the training data. This patch implements a mapping between the tokens
in the network output and the tokens that are actually present in the
training data. As a side effect, any reordering of the token list will
be applied as well.