Add skipFrame animation to AnimationQueue
Sometimes it is desireable to start an animation only after the current
frame has been fully processed and the call stack is empty. This is
especially important for CSS animations that start counting down their
duration immediately even if the current frame is a slow one due to some
heavy computations.