Reload app when resetting or idle to work around latency issue
For some yet unknown reason, the latency of the audio context increases
over time until the app becomes unusable. The only way to resolve this
seems to be to recreate the audio context and all audio nodes depending
on it. This would require to place an additional layer of
abstraction in between the app logic and all web audio related code.
Therefore, this commit implements a slower, but much simpler workaround
by reloading the app when it has been idle and when it is reset to the
recording screen.